Replace nodemon with onchange
nodemon didn't work for our purpose because it expects to restart a
server. we can't simply restart the server because we need to build the
client files too. This is now solved by providing a start:dev script
that lets you run the server if you want to anyway and for all other
purposes builds first then runs the build.